requirements.txt
opencv-python == 4.8.0.74
pillow == 10.0.0
pillow-avif-plugin == 1.3.1
pymongo == 4.4.*
python-dotenv == 1.0.0
scrapeops-scrapy == 0.5.2
scrapy == 2.10.0
scrapy-playwright >= 0.0.28